extracting API implementation from Fleet singleton. resolves Fleet/Device cycle.
