+++ packages/data-sdk/src/api/getPeers.ts
@@ -1,0 +1,16 @@
+import { IRtcPeer } from "@formant/realtime-sdk/dist/model/IRtcPeer";
+
+import { Authentication } from "../Authentication";
+import { defaultRtcClientPool } from "../AppRtcClientPools";
+
+export async function getPeers(): Promise<IRtcPeer[]> {
+  if (!Authentication.token) {
+    throw new Error("Not authenticated");
+  }
+  const rtcClient = defaultRtcClientPool.get();
+  try {
+    return await rtcClient.getPeers();
+  } finally {
+    await rtcClient.shutdown();
+  }
+}
