+++ packages/data-sdk/src/devices/ICustomDataChannelCreator.ts
@@ -1,0 +1,8 @@
+import { DataChannel } from "../DataChannel";
+
+export interface ICustomDataChannelCreator {
+  createCustomDataChannel(
+    channelName: string,
+    rtcConfig?: RTCDataChannelInit
+  ): Promise<DataChannel>;
+}
