+++ packages/data-sdk/src/model/TelemetryResult.ts
@@ -1,0 +1,7 @@
+export interface TelemetryResult {
+  deviceId: string;
+  name: string;
+  points: [number, any][];
+  tags: { [key in string]: string | number };
+  type: string;
+}
