--- packages/data-sdk/src/utils/index.ts
@@ -1,5 +1,0 @@
-export * from "./aggregateFunctionUtils";
-export * from "./numericAggregateUtils";
-export * from "./timeout";
-export * from "./serializeHash";
-export * from "./isRtcPeer";
